# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /www/portal/overview:
    get:
      operationId: getPortalOverview
      tags:
        - portal
      summary: 'Company overview - license key, live status, subscription details'
      responses:
        '200':
          description: Portal overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalOverview'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /www/portal/billing:
    post:
      operationId: createBillingSession
      tags:
        - portal
      summary: Create a Stripe billing portal session (manage/upgrade/cancel subscription)
      responses:
        '200':
          description: Billing portal session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSession'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /www/portal/users:
    get:
      operationId: listCompanyUsers
      tags:
        - portal
      summary: List company users and their roles
      responses:
        '200':
          description: Company user directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyUser'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: inviteCompanyUser
      tags:
        - portal
      summary: Invite a user by email with a role (returns a shareable invite link)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteInput'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResult'
        '400':
          description: Email already in company or over seat limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/www/portal/users/{userId}':
    patch:
      operationId: updateCompanyUser
      tags:
        - portal
      summary: Change a user's role or deactivate/reactivate them
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUserUpdate'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyUser'
        '400':
          description: Invalid change (e.g. demoting the master admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/www/portal/users/{userId}/invite-code':
    get:
      operationId: getInviteCode
      tags:
        - portal
      summary: View the pending invite code for an invited user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pending invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteCode'
        '403':
          description: Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or no longer pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: regenerateInviteCode
      tags:
        - portal
      summary: Regenerate the invite code for a still-pending invitee (invalidates the old code)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegeneratedInviteCode'
        '403':
          description: Requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or no longer pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
          description: Machine-readable error code
...
    CheckoutSession:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: URL to redirect the browser to
...
    LicenseSummary:
      type: object
      required:
        - key
        - status
        - statusReason
      properties:
        key:
          type: string
        status:
          type: string
          enum:
            - active
            - paused
            - cancelled
          description: Effective status derived from the subscription lifecycle
        statusReason:
          type: string
          description: Human-readable explanation of the current status
        activatedAt:
          type:
            - string
            - 'null'
          format: date-time
        lastVerifiedAt:
          type:
            - string
            - 'null'
          format: date-time
        paidThrough:
          type:
            - string
            - 'null'
          format: date-time
          description: End of the last paid billing period
    SubscriptionSummary:
      type: object
      required:
        - tier
        - tierName
        - seats
        - status
      properties:
        tier:
          type: string
        tierName:
          type: string
        seats:
          type:
            - integer
            - 'null'
          description: Seat limit for the tier; null means unlimited
        status:
          type: string
          description: 'Raw Stripe subscription status (active, past_due, canceled, ...)'
        currentPeriodEnd:
          type:
            - string
            - 'null'
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
        priceMonthlyCents:
          type:
            - integer
            - 'null'
    PortalOverview:
      type: object
      required:
        - company
        - license
        - subscription
        - seatsUsed
      properties:
        company:
          type: object
          required:
            - id
            - name
            - contactEmail
          properties:
            id:
              type: string
            name:
              type: string
            contactEmail:
              type: string
        license:
          $ref: '#/components/schemas/LicenseSummary'
        subscription:
          $ref: '#/components/schemas/SubscriptionSummary'
        seatsUsed:
          type: integer
    CompanyUser:
      type: object
      required:
        - id
        - email
        - name
        - role
        - active
        - isMasterAdmin
        - status
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum:
            - admin
            - manager
            - staff
            - readonly
        active:
          type: boolean
        isMasterAdmin:
          type: boolean
        username:
          type:
            - string
            - 'null'
          description: Admin-chosen desktop sign-in username; when null the desktop app derives one from the email local part
        status:
          type: string
          enum:
            - active
            - invited
            - deactivated
        createdAt:
          type:
            - string
            - 'null'
          format: date-time
    InviteInput:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - manager
            - staff
            - readonly
    InviteResult:
      type: object
      required:
        - user
        - inviteCode
        - inviteCodeExpiresAt
        - emailSent
      properties:
        user:
          $ref: '#/components/schemas/CompanyUser'
        inviteCode:
          type: string
          description: Single-use code the invitee enters in the desktop app to activate it and set up their account
        inviteCodeExpiresAt:
          type: string
          format: date-time
          description: When the invite code stops being redeemable (14 days after generation)
        emailSent:
          type: boolean
          description: Whether the invite email was sent to the invited user
    InviteCode:
      type: object
      required:
        - inviteCode
        - inviteCodeExpiresAt
        - email
        - role
      properties:
        inviteCode:
          type: string
        inviteCodeExpiresAt:
          type: string
          format: date-time
          description: When this code stops being redeemable (codes created before expiry tracking fall back to creation time + 14 days)
        email:
          type: string
        role:
          type: string
          enum:
            - admin
            - manager
            - staff
            - readonly
    RegeneratedInviteCode:
      allOf:
        - $ref: '#/components/schemas/InviteCode'
        - type: object
          required:
            - emailSent
          properties:
            emailSent:
              type: boolean
              description: Whether the fresh invite code was emailed to the invitee
    CompanyUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum:
            - admin
            - manager
            - staff
            - readonly
        active:
          type: boolean
        username:
          type:
            - string
            - 'null'
          maxLength: 32
          description: Desktop sign-in username (lowercase letters, digits, ".", "_", "-"); null clears it so the desktop app derives one from the email
...